requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
array-record==0.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dm-tree==0.1.8
dnspython==2.3.0
docutils==0.19
etils==1.2.0
Flask==2.2.3
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.0.4
multidict==6.0.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.2
opt-einsum==3.3.0
packaging==23.0
Pillow==9.5.0
promise==2.3
protobuf==3.20.3
psutil==5.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
requests-oauthlib==1.3.1
retry==0.9.2
rsa==4.9
scipy==1.10.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.9.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-macos==2.12.0
tensorflow-metadata==1.13.0
tensorflow-metal==0.8.0
termcolor==2.2.0
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.8.2
zipp==3.15.0
